pyproject.toml
[tool.poetry]
name = "parser"
version = "0.1.0"
description = "My description"
authors = ["Dm1trySereda <dima.sereda.d@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
redis = "^5.0.4"
sqlalchemy = "^2.0.30"
mysqlclient = "^2.2.4"
pydantic-settings = "^2.2.1"
mysql-connector-python = "^8.4.0"
aiomysql = "^0.2.0"
uvicorn = "^0.29.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
pydantic = "^2.7.1"
celery = "^5.4.0"
aiohttp = "^3.9.5"
bs4 = "^0.0.2"
lxml = "^5.2.2"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
numpy = "^1.26.4"


[tool.poetry.dev-dependencies]
alembic = "^1.13.1"
black = "^24.4.2"
mypy = "^1.10.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
flake8-variables-names = "^0.0.6"
pep8-naming = "^0.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"